requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
email-validator==2.2.0
itsdangerous==2.2.0
markdown2==2.4.12
bleach==6.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
